misc/HtmlTestSuite/results/test-0410.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSSOM Test: Table with border-collapse: collapse's clientWidth/Height</title>
<link href="https://www.w3.org/TR/CSS2/tables.html#model" rel="help"/>
<link href="https://www.w3.org/TR/CSS2/tables.html#collapsing-borders" rel="help"/>
<link href="https://drafts.csswg.org/cssom-view/#extension-to-the-element-interface" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  table {
    border-collapse: collapse;
  }
  td {
    border: 8px solid #faa;
  }
</style>
</head><body><table>
  <tbody>
    <tr>
      <td>a</td>
      <td>b</td>
    </tr>
    <tr>
      <td>c</td>
      <td>d</td>
    </tr>
  </tbody>
</table>
<script type="text/javascript">
  var table = document.getElementsByTagName('table')[0];
  test(function() {
    assert_equals(table.clientWidth, table.offsetWidth);
    assert_equals(table.clientHeight, table.offsetHeight);
  }, "Table's clientWidth/Height and OffsetWidth/Height should be the same");
</script>
</body></html>
